add npm packages
add npm packages for the
- language server
- language client for the monaco editor
- parser
- demo for the language server
- demo for the parser
