+++ languageserver/README.md
@@ -1,0 +1,9 @@
+# Cadence Language Server
+
+## Development
+
+### Building for WebAssembly
+
+```sh
+GOARCH=wasm GOOS=js go build -o languageserver.wasm  ./cmd/languageserver
+```
