--- languageserver/client/.gitignore
@@ -1,4 +1,0 @@
-.cache
-node_modules
-dist/*
-!dist/index.html
