+++ npm-packages/cadence-language-server-demo/.gitignore
@@ -1,0 +1,1 @@
+dist
