+++ npm-packages/cadence-language-server-demo/README.md
@@ -1,0 +1,7 @@
+# Cadence Language Server Demo
+
+## Running
+
+```sh
+npm run start
+```
