+++ npm-packages/cadence-language-server-demo/package.json
@@ -1,28 +1,32 @@
 {
-  "name": "monaco-cadence",
-  "version": "1.0.0",
+  "name": "cadence-language-server-demo",
+  "version": "0.8.0",
   "description": "An example client for the Cadence Language Server, based on Monaco",
   "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "watch": "webpack --watch",
     "start": "webpack-dev-server --open",
-    "build": "webpack"
+    "build": "webpack --mode=production"
   },
   "devDependencies": {
+    "copy-webpack-plugin": "^6.1.0",
     "css-loader": "^3.6.0",
     "file-loader": "^6.0.0",
+    "html-webpack-plugin": "^4.4.1",
     "monaco-editor-webpack-plugin": "^1.9.0",
     "style-loader": "^1.2.1",
-    "ts-loader": "^8.0.1",
-    "typescript": "^3.9.6",
-    "webpack": "^4.43.0",
+    "ts-loader": "^8.0.3",
+    "typescript": "^4.0.2",
+    "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
   },
   "dependencies": {
+    "cadence-language-server": "file:../cadence-language-server",
     "monaco-editor": "^0.21.0",
     "monaco-languageclient": "^0.13.0",
+    "monaco-languageclient-cadence": "file:../monaco-languageclient-cadence",
     "vscode-jsonrpc": "^5.0.1"
   }
 }
