+++ npm-packages/cadence-language-server-demo/src/index.html
@@ -35,7 +35,5 @@
     <button id="button2">Stop</button>
 </div>
 
-<script src="bundle.js"></script>
-
 </body>
 </html>
