+++ npm-packages/cadence-language-server/.gitignore
@@ -1,0 +1,1 @@
+dist
