+++ npm-packages/cadence-language-server/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "cadence-language-server",
+  "version": "0.8.0",
+  "description": "The Cadence Language Server",
+  "homepage": "https://github.com/onflow/cadence",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/onflow/cadence.git"
+  },
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "tsc"
+  },
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "typescript": "^4.0.2"
+  },
+  "dependencies": {
+    "vscode-jsonrpc": "^5.0.1"
+  },
+  "files": [
+    "dist/**/*"
+  ]
+}
