+++ npm-packages/cadence-language-server/src/index.ts
@@ -17,12 +17,7 @@
  */
 
 import {Message} from "vscode-jsonrpc/lib/messages";
-
-// The global `Go` is declared by `wasm_exec.js`.
-// Instead of improving the that file, we use it as-is,
-// because it is maintained by the Go team and
-
-declare var Go: any
+import {go} from './go.js'
 
 // Callbacks defines the functions that the language server calls
 // and that need to be implemented by the client.
@@ -31,45 +26,126 @@ export interface Callbacks {
   // The function that the language server calls
   // to write a message object to the client.
   // The object is a JSON-RPC request/response object
-  toClient(message: Message): void
+  toClient?(message: Message): void
 
   // The function that the language server calls
   // to get the code for an imported address, if any
-  getAddressCode(address: string): string | undefined
+  getAddressCode?(address: string): string | undefined
 
   // The function that the language client calls
   // to write a message object to the server.
   // The object is an JSON-RPC request/response object
-  toServer(error: any, message: Message): void
+  toServer?(error: any, message: Message): void
 
   // The function that the language client can call
   // to notify the server that the client is closing
-  onClientClose(): void
+  onClientClose?(): void
 
   // The function that the language server can call
   // to notify the client that the server is closing
-  onServerClose(): void
+  onServerClose?(): void
 }
 
 declare global {
   interface Window {
-    [index: string]: any;
+    [key: string]: any
   }
 }
 
 export class CadenceLanguageServer {
 
-  static isLoaded = false
+  private static functionNamePrefix = "CADENCE_LANGUAGE_SERVER"
+  private static loaded = false
 
-  private static async load() {
-    if (this.isLoaded) {
+  private static functionName(name: string): string {
+    return `__${CadenceLanguageServer.functionNamePrefix}_${name}__`
+  }
+
+  private functionName(name: string): string {
+    return `__${CadenceLanguageServer.functionNamePrefix}_${this.id}_${name}__`
+  }
+
+  public readonly callbacks: Callbacks
+  public readonly id: number
+  private isClientClosed: boolean = false
+
+  public static async create(binaryLocation: string, callbacks: Callbacks): Promise<CadenceLanguageServer> {
+    await this.ensureLoaded(binaryLocation)
+    return new CadenceLanguageServer(callbacks)
+  }
+
+  private static async ensureLoaded(binaryLocation: string) {
+    if (this.loaded) {
       return
     }
 
-    const wasm = await fetch("./languageserver.wasm")
-    const go = new Go()
-    const module = await WebAssembly.instantiateStreaming(wasm, go.importObject)
+    this.setWriteSync()
 
+    await this.load(binaryLocation)
+    this.loaded = true
+  }
+
+  private static async load(binaryLocation: string) {
+    const binary = fetch(binaryLocation)
+    const instantiatedSource = await WebAssembly.instantiateStreaming(binary, go.importObject)
+    // NOTE: don't await the promise, just ignore it, as it is only resolved when the program exists
+    go.run(instantiatedSource.instance).then(() => {})
+  }
+
+  private constructor(callbacks: Callbacks) {
+    this.callbacks = callbacks
+
+    // The language server, written in Go and compiled to WebAssembly, interacts with this JS environment
+    // by calling global functions. There does not seem to be support yet to directly import functions
+    // from the JS environment into the WebAssembly environment
+
+    this.id = window[CadenceLanguageServer.functionName('start')]()
+
+    window[this.functionName('toClient')] = (message: string): void => {
+      if (!callbacks.toClient) {
+        return
+      }
+      callbacks.toClient(JSON.parse(message))
+    }
+
+    window[this.functionName('getAddressCode')] = (address: string): string | undefined => {
+      if (!callbacks.getAddressCode) {
+        return undefined
+      }
+
+      return callbacks.getAddressCode(address)
+    }
+
+    window[this.functionName('onServerClose')] = (): void => {
+      if (!callbacks.onServerClose) {
+        return
+      }
+      callbacks.onServerClose()
+    }
+
+    callbacks.toServer = (error: any, message: any) => {
+      window[this.functionName('toServer')](error, JSON.stringify(message))
+    }
+
+    callbacks.onClientClose = () => {
+      if (this.isClientClosed) {
+        return
+      }
+      this.isClientClosed = true
+      window[this.functionName('onClientClose')]()
+    }
+  }
+
+  close() {
+    const { onClientClose } = this.callbacks;
+    if (!onClientClose) {
+      return
+    }
+    onClientClose()
+  }
+
+  // setWriteSync installs the writeSync filesystem handler that the Go WebAssembly binary calls
+  private static setWriteSync() {
     // For each file descriptor, buffer the written content until reaching a newline
 
     const outputBuffers = new Map<number, string>()
@@ -104,70 +180,6 @@ export class CadenceLanguageServer {
       outputBuffers.set(fileDescriptor, outputBuffer)
 
       return buf.length
-    }
-
-    go.run(module.instance)
-
-    this.isLoaded = true
-  }
-
-  private static functionNamePrefix = "CADENCE_LANGUAGE_SERVER"
-
-  private static functionName(name: string): string {
-    return `__${CadenceLanguageServer.functionNamePrefix}_${name}__`
-  }
-
-  private functionName(name: string): string {
-    return `__${CadenceLanguageServer.functionNamePrefix}_${this.id}_${name}__`
-  }
-
-  static async create(callbacks: Callbacks): Promise<CadenceLanguageServer> {
-
-    await this.load()
-
-    return new CadenceLanguageServer(callbacks)
-  }
-
-  public readonly id: number
-  private isClientClosed: boolean
-
-  private constructor(callbacks: Callbacks) {
-
-    // The language server, written in Go and compiled to WebAssembly, interacts with this JS environment
-    // by calling global functions. There does not seem to be support yet to directly import functions
-    // from the JS environment into the WebAssembly environment
-
-    this.id = window[CadenceLanguageServer.functionName('start')]()
-
-    window[this.functionName('toClient')] = (message: string): void => {
-      callbacks.toClient(JSON.parse(message))
-    }
-
-    window[this.functionName('getAddressCode')] = (address: string): string | undefined => {
-      if (!callbacks.getAddressCode) {
-        return undefined
-      }
-
-      return callbacks.getAddressCode(address)
-    }
-
-    window[this.functionName('onServerClose')] = (): void => {
-      if (!callbacks.onServerClose) {
-        return
-      }
-      callbacks.onServerClose()
-    }
-
-    callbacks.toServer = (error: any, message: any) => {
-      window[this.functionName('toServer')](error, JSON.stringify(message))
-    }
-
-    callbacks.onClientClose = () => {
-      if (this.isClientClosed) {
-        return
-      }
-      this.isClientClosed = true
-      window[this.functionName('onClientClose')]()
     }
   }
 }
