+++ npm-packages/cadence-parser-demo/README.md
@@ -1,0 +1,7 @@
+# Cadence Parser Demo
+
+## Running
+
+```sh
+npm run start
+```
