+++ npm-packages/cadence-parser-demo/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "cadence-parser-demo",
+  "version": "0.8.0",
+  "description": "A demo for the Cadence parser",
+  "private": true,
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "watch": "webpack --watch",
+    "start": "webpack-dev-server --open",
+    "build": "webpack --mode=production"
+  },
+  "devDependencies": {
+    "@types/react-dom": "^16.9.8",
+    "copy-webpack-plugin": "^6.1.0",
+    "css-loader": "^3.6.0",
+    "file-loader": "^6.0.0",
+    "html-webpack-plugin": "^4.4.1",
+    "monaco-editor-webpack-plugin": "^1.9.0",
+    "style-loader": "^1.2.1",
+    "ts-loader": "^8.0.3",
+    "typescript": "^4.0.2",
+    "webpack": "^4.44.2",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
+  },
+  "dependencies": {
+    "cadence-parser": "file:../cadence-parser",
+    "monaco-editor": "^0.20.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-json-view": "^1.19.1"
+  }
+}
