+++ npm-packages/cadence-parser-demo/src/index.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Cadence Parser Example</title>
+    <style>
+        body {
+            font-family: monospace;
+            display: flex;
+            flex-direction: column;
+            align-items: stretch;
+            min-height: 96vh;
+        }
+
+        #panels {
+            display: flex;
+            align-items: stretch;
+            flex: 100%;
+        }
+
+        #editor {
+            border: 1px solid #ccc;
+            flex: 50%;
+        }
+
+        #ast {
+            flex: 50%;
+        }
+    </style>
+</head>
+<body>
+
+<h1>Cadence Parser Demo</h1>
+
+<div id="panels">
+    <div id="editor"></div>
+    <div id="ast"></div>
+</div>
+
+</body>
+</html>
