+++ npm-packages/cadence-parser/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "cadence-parser",
+  "version": "0.8.0",
+  "description": "The Cadence Language Server",
+  "homepage": "https://github.com/onflow/cadence",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/onflow/cadence.git"
+  },
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "tsc"
+  },
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "typescript": "^4.0.2"
+  },
+  "files": [
+    "dist/**/*"
+  ]
+}
