+++ npm-packages/monaco-languageclient-cadence/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "monaco-languageclient-cadence",
+  "version": "0.1.0",
+  "description": "The Cadence Language Server",
+  "homepage": "https://github.com/onflow/cadence",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/onflow/cadence.git"
+  },
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "tsc"
+  },
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "typescript": "^4.0.2"
+  },
+  "dependencies": {
+    "cadence-language-server": "file:../cadence-language-server",
+    "monaco-languageclient": "^0.13.0",
+    "vscode-jsonrpc": "^5.0.1"
+  },
+  "files": [
+    "dist/**/*"
+  ]
+}
